Fix errors.

+++ cache/shard.go
@@ -1,14 +1,14 @@
 package cache
 
-func main_shard_server(shard_index int) RedisServer {
+func main_shard_server(shard_index uint64) RedisServer {
 	return shard_server(shard_index, RedisSettings.MainServers)
 }
 
-func old_shard_server(shard_index int) RedisServer {
+func old_shard_server(shard_index uint64) RedisServer {
 	return shard_server(shard_index, RedisSettings.OldServers)
 }
 
-func shard_server(shard_index int, servers []RedisServer) RedisServer {
-	server_index := shard_index % len(servers)
+func shard_server(shard_index uint64, servers []RedisServer) RedisServer {
+	server_index := shard_index % uint64(len(servers))
 	return servers[server_index]
 }
